Use Nixpkgs to compose python packages.
The Python rules don't have a `py_import` rule analogous to
`cc_import` or `haskell_import`. This means that we can't in a `BUILD`
file directly reference Python libraries that are installed in some
system location. The workaround is to ask Nix to provide us with
a Python interpreter that has all the libraries we want in-scope. This
is achieved using the `python27Packages.withPython` function,
documented [here](https://nixos.org/nixpkgs/manual/#using-python).